requirements.txt
aiogram>=3.10,<4
python-dotenv>=1.0,<2
